third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-script-in-iframe-expected.txt

CONSOLE WARNING: line 1: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/security/mixedContent/resources/script.js': this content should also be loaded over HTTPS.

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/security/mixedContent/resources/script.js': this content should also be loaded over HTTPS.

This test loads a secure iframe that loads an insecure script. We should trigger a mixed content callback even though the main frame is HTTP because the HTTPS frame's origin is contaminated with an insecure script.
